Refactoring to address a couple of PR comments
